pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tigre"
dynamic = ["version"]
description = "tigre - Tool for InterGenic Region Extraction"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"

authors = [
    { name = "Breno Dupin", email = "breno.dupin@gmail.com" },
    { name = "Matheus Sanita Lima", email = "msanital@uwo.ca" },
]

dependencies = [
    "pandas>=1.5.3,<3.0.0",
]

[project.optional-dependencies]
bio = ["biopython>=1.80"]
all = ["tigre[bio]"]

[project.urls]
Homepage = "https://github.com/brenodupin/tigre"
Source = "https://github.com/brenodupin/tigre"

[project.scripts]
tigre = "tigre.cli:cli_entrypoint"

[tool.hatch.version]
path = "src/tigre/__init__.py"

[tool.black]
line-length = 88
